divide lib into more modules
